Refactor ConnectionBuddyConfiguration
Field 'context' renamed to 'applicationContext', as it is, in fact, an application context. Getter unchanged, because it is public API, but added docstring.
Added a 'wifiManager' field for easier access to the wifi manager service through the configuration.
It, as well as the 'connectivityManager' field, is marked with the '@NonNull' annotation, and is null-checked before assignment.